fix: keep unmounted workspace portals hidden
