+++ cmuxTests/GhosttyTerminalViewVisibilityPolicyTests.swift
@@ -1,0 +1,60 @@
+import XCTest
+
+#if canImport(cmux_DEV)
+@testable import cmux_DEV
+#elseif canImport(cmux)
+@testable import cmux
+#endif
+
+final class GhosttyTerminalViewVisibilityPolicyTests: XCTestCase {
+    func testImmediateStateUpdateAllowedWhenDesiredStateIsHidden() {
+        XCTAssertTrue(
+            GhosttyTerminalView.shouldApplyImmediateHostedStateUpdate(
+                desiredVisibleInUI: false,
+                hostedViewHasSuperview: true,
+                isBoundToCurrentHost: false
+            )
+        )
+    }
+
+    func testImmediateStateUpdateAllowedWhenBoundToCurrentHost() {
+        XCTAssertTrue(
+            GhosttyTerminalView.shouldApplyImmediateHostedStateUpdate(
+                desiredVisibleInUI: true,
+                hostedViewHasSuperview: true,
+                isBoundToCurrentHost: true
+            )
+        )
+    }
+
+    func testImmediateStateUpdateSkippedForStaleHostBoundElsewhere() {
+        XCTAssertFalse(
+            GhosttyTerminalView.shouldApplyImmediateHostedStateUpdate(
+                desiredVisibleInUI: true,
+                hostedViewHasSuperview: true,
+                isBoundToCurrentHost: false
+            )
+        )
+    }
+
+    func testImmediateStateUpdateAllowedWhenUnboundAndNotAttachedAnywhere() {
+        XCTAssertTrue(
+            GhosttyTerminalView.shouldApplyImmediateHostedStateUpdate(
+                desiredVisibleInUI: true,
+                hostedViewHasSuperview: false,
+                isBoundToCurrentHost: false
+            )
+        )
+    }
+
+    func testInteractiveGeometryResizeUsesImmediatePortalSyncDecision() {
+        XCTAssertTrue(
+            GhosttyTerminalView.shouldSynchronizePortalGeometryImmediately(
+                hostInLiveResize: false,
+                windowInLiveResize: false,
+                interactiveGeometryResizeActive: true
+            ),
+            "Interactive resize should use the immediate portal sync path"
+        )
+    }
+}
